django-site/projet/api/requirements.txt
Django==4.2.15
djangorestframework==3.14.0
psycopg2-binary==2.9.6  # PostgreSQL driver for Django
